Create 2024-08-13-ComputingScience.md
What is the difference between Computer Science and Computing Science?

Computer science is fundamentally the “know-how” of a computer: a machine that “is able to solve problems.” 

Computing science is the knowing of how to compute : that is, how to “put” things together: (com-put-e).

It’s the wisdom of groupings, of algorithms, of data structures, the understanding that plants compute how to grow, based on limited resources, that people decide how to shop, based on what’s available to them, and how we all make the decisions that shape our day-to-day. 


I will define computing science as a primarily present-and-future-based discipline. The art of : what questions will need to be solved; and, how do we solve them?